requirements.txt
Flask==1.0.2
requests==2.22.0
Flask-Session==0.3
google-auth==1.6.2
requests_oauthlib==1.3.0
google-cloud-datastore==1.7.3